third_party/WebKit/LayoutTests/fast/scroll-behavior/scroll-customization/scrollstate-distribute-to-scroll-chain-descendant.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>ScrollState constructor behaves correctly</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
* {
  margin: 0 0;
}
div {
  width: 400px;
  height: 400px;
  scroll-customization: auto;
  overflow: scroll;
}
</style>

<p> Verify the correctness of distributeToScrollChainDescendant(). </p>

<script>
  test(function() {
    assert_true('ScrollState' in window, "'ScrollState' in window");
    // TODO(tdresser): Don't rely on window.internals. See crbug.com/483091 for
    // details.
    assert_true(window.internals !== null, "'internals' in window");
  }, "These tests only work with scroll customization enabled.");

  var elementCount = 10;
  var remainingNumberOfTimesToBeCalled = elementCount;
  var elements = [];

  var distributeScroll = function(scrollState) {
    this.calledOrder = elementCount - remainingNumberOfTimesToBeCalled;
    remainingNumberOfTimesToBeCalled--;
    scrollState.distributeToScrollChainDescendant();
  }

  var parentNode = document.body;
  for (var i = 0; i < elementCount; ++i) {
    var element = document.createElement("div");
    element.creationOrder = i;
    element.id = 'e' + i;
    element.style.scrollCustomization = 'auto';
    parentNode.appendChild(element);
    elements.push(element);
    parentNode = element;
  }

  // Normal codepath for scrolling involves notifying all the elements in the
  // scroll chain about a GestureScrollBegin. The element finds the direction
  // of the scroll and decides whether or not its customization matches the
  // direction. Since this test uses a synthetic chain, this step needs to
  // be simulated by a GestureScrollBegin. This will also call an related
  // customized scroll handlers. Therefore, to avoid the unwanted calls to the
  // corresponding distributeScroll handlers, they are set up after the
  // gesture.
  eventSender.gestureScrollBegin(200, 200);

  elements.forEach((el) => {
    el.setDistributeScroll(distributeScroll.bind(el),
     "disable-native-scroll");
  });


  test(() => {
    var scrollState = new ScrollState(
        { deltaX: 100, isDirectManipulation: true});
    internals.setScrollChain(scrollState, elements);
    scrollState.distributeToScrollChainDescendant();
    assert_equals(remainingNumberOfTimesToBeCalled, 0);
    for (var i = 0; i < elementCount; ++i) {
      assert_equals(elements[i].creationOrder, elements[i].calledOrder);
    }
  }, "distributeToScrollChainDescendant propagates correctly");

  test(() => {
    var scrollState =
        new ScrollState({deltaX: 100, isDirectManipulation: true});
    internals.setScrollChain(scrollState, []);
    assert_equals(0, remainingNumberOfTimesToBeCalled);
    scrollState.distributeToScrollChainDescendant();
    assert_equals(0, remainingNumberOfTimesToBeCalled);
  },
  "distributeToScrollChainDescendant with empty scroll chain does nothing.");
</script>
